requirements.txt
# Core web framework
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Document processing
PyMuPDF==1.23.22
pdfplumber==0.10.4
pypdf2==3.0.1
python-docx==1.1.0

# Text processing and embedding
langchain==0.1.6
sentence-transformers==2.2.2
transformers>=4.36.2
torch>=2.1.2
accelerate>=0.20.0

# Vector store
faiss-cpu==1.7.4

# Utilities
python-dotenv==1.0.0
requests==2.31.0
numpy>=1.24.4,<2.0.0

# Optional: For better tokenization
tiktoken==0.5.2